requirements.txt
aiodns==3.0.0
aiofile==3.7.4
aiohttp==3.8.1
aiosignal==1.2.0
appdirs==1.4.4
appnope==0.1.2
asgiref==3.4.1
astroid==2.9.3
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.10.0
black==21.12b0
boto3==1.20.39
botocore==1.23.39
CacheControl==0.12.10
cachetools==4.2.4
caio==0.9.3
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.10
click==8.0.3
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.1
Django==3.2.11
django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers==3.11.0
django-crontab==0.7.1
django-environ==0.8.1
django-model-utils==4.2.0
django-otp==1.1.3
django-user-visit==0.5.1
djangorestframework==3.13.1
docutils==0.18.1
drf-yasg==1.20.0
executing==0.8.2
fcm-django==1.0.8
firebase-admin==5.2.0
frozenlist==1.3.0
google-api-core==2.4.0
google-api-python-client==2.36.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-cloud-core==2.2.2
google-cloud-firestore==2.3.4
google-cloud-storage==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
grpcio==1.43.0
grpcio-status==1.43.0
gunicorn==20.1.0
httplib2==0.20.2
idna==3.3
inflection==0.5.1
install==1.3.5
ipython==8.0.1
isort==5.10.1
itypes==1.2.0
jedi==0.18.1
Jinja2==3.0.3
jmespath==0.10.0
lazy-object-proxy==1.7.1
linkpreview==0.2.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mccabe==0.6.1
msgpack==1.0.3
multidict==5.2.0
mypy-extensions==0.4.3
oauthlib==3.1.1
packaging==21.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.0
platformdirs==2.4.1
prompt-toolkit==3.0.24
proto-plus==1.19.8
protobuf==3.19.3
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==4.1.2
pycodestyle==2.8.0
pycparser==2.21
pycryptodomex==3.12.0
pyfcm==1.5.4
Pygments==2.11.2
PyJWT==2.3.0
pylint==2.12.2
pylint-plugin-utils==0.7
pyparsing==3.0.6
python-dateutil==2.8.2
python-memcached==1.59
pytz==2021.3
regex==2022.1.18
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
s3transfer==0.5.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
stack-data==0.1.4
stream-chat==4.0.0
tokenizer==3.3.2
toml==0.10.2
tomli==1.2.3
traitlets==5.1.1
twilio==7.5.0
typed-ast==1.5.1
typing_extensions==4.0.1
ua-parser==0.10.0
uritemplate==4.1.1
urllib3==1.26.8
user-agents==2.2.0
wcwidth==0.2.5
wrapt==1.13.3
yarl==1.7.2
